+++ .idea/misc.xml
@@ -5,7 +5,7 @@
       <map>
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.31100006103515626" />
         <entry key="app/src/main/res/layout/content_main.xml" value="0.22" />
-        <entry key="app/src/main/res/layout/dashboard_fragment.xml" value="0.18" />
+        <entry key="app/src/main/res/layout/dashboard_fragment.xml" value="0.33" />
         <entry key="app/src/main/res/layout/fragment_first.xml" value="0.22" />
         <entry key="app/src/main/res/layout/fragment_second.xml" value="0.25" />
       </map>
