--- app/src/main/java/com/masgray/findfriends/MainActivity.kt
@@ -1,29 +1,0 @@
-package com.masgray.findfriends
-
-import android.content.Intent
-import android.os.Bundle
-import com.google.android.material.snackbar.Snackbar
-import androidx.appcompat.app.AppCompatActivity
-import androidx.navigation.findNavController
-import androidx.navigation.ui.AppBarConfiguration
-import androidx.navigation.ui.navigateUp
-import androidx.navigation.ui.setupActionBarWithNavController
-import android.view.Menu
-import android.view.MenuItem
-import android.widget.Button
-import com.masgray.findfriends.databinding.ActivityMainBinding
-
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        val findfriendbtn = findViewById<Button>(R.id.findFriendButton)
-        findfriendbtn.setOnClickListener {
-            val intent = Intent(this, FindFriendsActivity::class.java)
-            startActivity(intent)
-        }
-    }
-
-}
