+++ app/src/main/java/com/masgray/findfriends/ui/friendsList/FriendsListFragment.kt
@@ -1,0 +1,33 @@
+package com.masgray.findfriends.ui.friendsList
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+
+/**
+ * A simple [Fragment] subclass as the second destination in the navigation.
+ */
+class FriendsListFragment : Fragment() {
+
+
+    // This property is only valid between onCreateView and
+    // onDestroyView.
+
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View? {
+        return null
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+    }
+}
