+++ app/src/main/java/com/masgray/findfriends/ui/homeScreen/HomeScreenFragment.kt
@@ -1,19 +1,20 @@
-package com.masgray.findfriends
+package com.masgray.findfriends.ui.homeScreen
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
-import com.masgray.findfriends.databinding.FragmentFirstBinding
+import com.masgray.findfriends.R
+import com.masgray.findfriends.databinding.FragmentHomeScreenBinding
 
 /**
  * A simple [Fragment] subclass as the default destination in the navigation.
  */
-class FirstFragment : Fragment() {
+class HomeScreenFragment : Fragment() {
 
-    private var _binding: FragmentFirstBinding? = null
+    private var _binding: FragmentHomeScreenBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -24,7 +25,7 @@ class FirstFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
 
-        _binding = FragmentFirstBinding.inflate(inflater, container, false)
+        _binding = FragmentHomeScreenBinding.inflate(inflater, container, false)
         return binding.root
 
     }
@@ -33,7 +34,7 @@ class FirstFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
 
         binding.buttonFirst.setOnClickListener {
-            findNavController().navigate(R.id.action_FirstFragment_to_SecondFragment)
+            findNavController().navigate(R.id.action_HomeScreenFragment_to_FindFriendsFragment)
         }
     }
