+++ app/src/main/res/navigation/nav_graph.xml
@@ -3,26 +3,25 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/FirstFragment">
+    app:startDestination="@id/HomeScreenFragment">
 
     <fragment
-        android:id="@+id/FirstFragment"
-        android:name="com.masgray.findfriends.FirstFragment"
-        android:label="@string/first_fragment_label"
-        tools:layout="@layout/fragment_first">
+        android:id="@+id/HomeScreenFragment"
+        android:name="com.masgray.findfriends.ui.homeScreen.HomeScreenFragment"
+        android:label="@string/home_screen_label"
+        tools:layout="@layout/fragment_home_screen">
 
         <action
-            android:id="@+id/action_FirstFragment_to_SecondFragment"
-            app:destination="@id/SecondFragment" />
+            android:id="@+id/action_HomeScreenFragment_to_FindFriendsFragment"
+            app:destination="@id/FriendsListFragment" />
     </fragment>
     <fragment
-        android:id="@+id/SecondFragment"
-        android:name="com.masgray.findfriends.SecondFragment"
-        android:label="@string/second_fragment_label"
-        tools:layout="@layout/fragment_second">
+        android:id="@+id/FriendsListFragment"
+        android:name="com.masgray.findfriends.ui.friendsList.FriendsListFragment"
+        android:label="@string/find_friends_fragment_label">
 
         <action
-            android:id="@+id/action_SecondFragment_to_FirstFragment"
-            app:destination="@id/FirstFragment" />
+            android:id="@+id/action_FriendsListFragment_to_HomeScreenFragment"
+            app:destination="@id/HomeScreenFragment" />
     </fragment>
 </navigation>
