+++ app/src/main/res/values/strings.xml
@@ -2,8 +2,8 @@
     <string name="app_name">FindFriends</string>
     <string name="action_settings">Settings</string>
     <!-- Strings used for fragments for navigation -->
-    <string name="first_fragment_label">First Fragment</string>
-    <string name="second_fragment_label">Second Fragment</string>
+    <string name="home_screen_label">Home Screen Fragment</string>
+    <string name="find_friends_fragment_label">Find Friends Fragment</string>
     <string name="next">Next</string>
     <string name="previous">Previous</string>
