+++ settings.gradle
@@ -14,3 +14,5 @@ dependencyResolutionManagement {
 }
 rootProject.name = "FindFriends"
 include ':app'
+include ':api'
+include ':feature'
